requirements.txt
alabaster==0.7.12
allennlp==0.8.3
appnope==0.1.0
args==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
aws-sam-translator==1.11.0
aws-xray-sdk==2.4.2
awscli==1.16.168
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.158
botocore==1.12.158
certifi==2019.3.9
cffi==1.12.3
cfn-lint==0.21.3
chardet==3.0.4
Click==7.0
clint==0.5.1
colorama==0.3.9
conllu==0.11
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.5
docker==4.0.1
docutils==0.14
ecdsa==0.13.2
editdistance==0.5.3
entrypoints==0.3
flaky==3.5.3
Flask==1.0.3
Flask-Cors==3.0.7
ftfy==5.5.1
future==0.17.1
gensim==3.6.0
gevent==1.4.0
greenlet==0.4.15
h5py==2.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsondiff==1.1.2
jsonnet==0.12.1
jsonpatch==1.23
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
moto==1.3.8
msgpack==0.5.6
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.1
notebook==5.7.8
numpy==1.16.3
numpydoc==0.9.1
overrides==1.9
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.0
pyparsing==2.4.0
pyrsistent==0.15.1
pytest==4.5.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-jose==3.0.1
pytorch-pretrained-bert==0.6.2
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
regex==2019.5.25
requests==2.21.0
responses==0.10.6
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
smart-open==1.8.4
snowballstemmer==1.2.1
spacy==2.1.4
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.0.5
tensorboardX==1.7
terminado==0.8.2
testpath==0.4.2
thinc==7.0.4
torch==1.1.0
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.4.2
word2number==1.1
wrapt==1.11.1
xmltodict==0.12.0
zipp==0.5.1
